+++ elasticsearch5/search-index/src/main/java/org/vertexium/elasticsearch5/ElasticsearchSearchIndexConfiguration.java
@@ -59,15 +59,15 @@ public class ElasticsearchSearchIndexConfiguration {
     public static final Integer LOG_REQUEST_SIZE_LIMIT_DEFAULT = null;
     public static final String MAX_QUERY_STRING_TERMS = "maxQueryStringTerms";
     public static final int MAX_QUERY_STRING_TERMS_DEFAULT = 100;
-    public static final String BULK_CORE_POOL_SIZE = "bulk.corePoolSize";
-    public static final String BULK_MAX_POOL_SIZE = "bulk.maxPoolSize";
+    public static final String BULK_POOL_SIZE = "bulk.poolSize";
+    public static final String BULK_BACKLOG_SIZE = "bulk.backlogSize";
     public static final String BULK_MAX_BATCH_SIZE = "bulk.maxBatchSize";
     public static final String BULK_MAX_BATCH_SIZE_IN_BYTES = "bulk.maxBatchSizeInBytes";
     public static final String BULK_BATCH_WINDOW_TIME = "bulk.batchWindowTime";
     public static final String BULK_MAX_FAIL_COUNT = "bulk.maxFailCount";
     public static final String BULK_REQUEST_TIMEOUT = "bulk.requestTimeout";
-    public static final int BULK_CORE_POOL_SIZE_DEFAULT = BulkUpdateServiceConfiguration.CORE_POOL_SIZE_DEFAULT;
-    public static final int BULK_MAX_POOL_SIZE_DEFAULT = BulkUpdateServiceConfiguration.MAX_POOL_SIZE_DEFAULT;
+    public static final int BULK_POOL_SIZE_DEFAULT = BulkUpdateServiceConfiguration.POOL_SIZE_DEFAULT;
+    public static final int BULK_BACKLOG_SIZE_DEFAULT = BulkUpdateServiceConfiguration.BACKLOG_SIZE_DEFAULT;
     public static final int BULK_MAX_BATCH_SIZE_DEFAULT = BulkUpdateServiceConfiguration.MAX_BATCH_SIZE_DEFAULT;
     public static final int BULK_MAX_BATCH_SIZE_IN_BYTES_DEFAULT = BulkUpdateServiceConfiguration.MAX_BATCH_SIZE_IN_BYTES_DEFAULT;
     public static final Duration BULK_BATCH_WINDOW_TIME_DEFAULT = BulkUpdateServiceConfiguration.BATCH_WINDOW_TIME_DEFAULT;
@@ -249,12 +249,12 @@ public class ElasticsearchSearchIndexConfiguration {
         return graphConfiguration.getInt(GraphConfiguration.SEARCH_INDEX_PROP_PREFIX + "." + MAX_QUERY_STRING_TERMS, MAX_QUERY_STRING_TERMS_DEFAULT);
     }
 
-    public int getBulkCorePoolSize() {
-        return graphConfiguration.getInt(GraphConfiguration.SEARCH_INDEX_PROP_PREFIX + "." + BULK_CORE_POOL_SIZE, BULK_CORE_POOL_SIZE_DEFAULT);
+    public int getBulkPoolSize() {
+        return graphConfiguration.getInt(GraphConfiguration.SEARCH_INDEX_PROP_PREFIX + "." + BULK_POOL_SIZE, BULK_POOL_SIZE_DEFAULT);
     }
 
-    public int getBulkMaxPoolSize() {
-        return graphConfiguration.getInt(GraphConfiguration.SEARCH_INDEX_PROP_PREFIX + "." + BULK_MAX_POOL_SIZE, BULK_MAX_POOL_SIZE_DEFAULT);
+    public int getBulkBacklogSize() {
+        return graphConfiguration.getInt(GraphConfiguration.SEARCH_INDEX_PROP_PREFIX + "." + BULK_BACKLOG_SIZE, BULK_BACKLOG_SIZE_DEFAULT);
     }
 
     public int getBulkMaxBatchSize() {