+++ elasticsearch7/search-index/src/test/java/org/vertexium/elasticsearch7/Elasticsearch7SearchIndexTest.java
@@ -422,6 +422,33 @@ public class Elasticsearch7SearchIndexTest extends GraphTestBase {
         }
     }
 
+    @Test
+    public void testManyWritesToSameElement() throws InterruptedException {
+        int threadCount = 10;
+        int numberOfTimerToWrite = 100;
+
+        Thread[] threads = new Thread[threadCount];
+        for (int i = 0; i < threads.length; i++) {
+            threads[i] = new Thread(() -> {
+                for (int write = 0; write < numberOfTimerToWrite; write++) {
+                    String keyAndValue = Thread.currentThread().getId() + "-" + write;
+                    getGraph().prepareVertex("v1", VISIBILITY_EMPTY)
+                        .addPropertyValue(keyAndValue, "name", keyAndValue, VISIBILITY_EMPTY)
+                        .save(AUTHORIZATIONS_EMPTY);
+                    getGraph().flush();
+                }
+            });
+        }
+        for (Thread thread : threads) {
+            thread.start();
+        }
+        for (Thread thread : threads) {
+            thread.join();
+        }
+        Vertex v1 = getGraph().getVertex("v1", AUTHORIZATIONS_EMPTY);
+        assertEquals(threadCount * numberOfTimerToWrite, count(v1.getProperties("name")));
+    }
+
     private long getNumQueries() {
         NodesStatsResponse nodeStats = new NodesStatsRequestBuilder(getSearchIndex().getClient(), NodesStatsAction.INSTANCE).get();